feat(client): add Upstash rate limiting to middleware
