+++ client/proxy.ts
@@ -1,25 +1,46 @@
-import { NextResponse } from "next/server";
-import type { NextRequest } from "next/server";
+import { NextResponse } from 'next/server';
+import type { NextRequest } from 'next/server';
+import { authRatelimit, apiRatelimit } from '@/lib/rate-limit';
 
-// Routes that require authentication
-const protectedRoutes = ["/account", "/orders"];
+const protectedRoutes = ['/account', '/orders'];
+
+function getIp(req: NextRequest): string {
+  return req.headers.get('x-forwarded-for')?.split(',')[0].trim() ?? '127.0.0.1';
+}
 
 export async function middleware(request: NextRequest) {
   const { pathname } = request.nextUrl;
+  const ip = getIp(request);
 
-  // Check if the route requires authentication
-  const isProtectedRoute = protectedRoutes.some((route) =>
-    pathname.startsWith(route)
-  );
+  // Rate limit auth endpoints
+  if (pathname.startsWith('/api/auth')) {
+    const { success, reset } = await authRatelimit.limit(ip);
+    if (!success) {
+      return NextResponse.json(
+        { error: 'Too many requests' },
+        { status: 429, headers: { 'Retry-After': String(Math.ceil((reset - Date.now()) / 1000)) } }
+      );
+    }
+  }
 
+  // Rate limit public API endpoints
+  if (pathname.startsWith('/api/')) {
+    const { success, reset } = await apiRatelimit.limit(ip);
+    if (!success) {
+      return NextResponse.json(
+        { error: 'Too many requests' },
+        { status: 429, headers: { 'Retry-After': String(Math.ceil((reset - Date.now()) / 1000)) } }
+      );
+    }
+  }
+
+  // Auth guard for protected routes
+  const isProtectedRoute = protectedRoutes.some((r) => pathname.startsWith(r));
   if (isProtectedRoute) {
-    // Get session from cookie
-    const sessionToken = request.cookies.get("better-auth.session_token")?.value;
-
+    const sessionToken = request.cookies.get('better-auth.session_token')?.value;
     if (!sessionToken) {
-      // Redirect to login if not authenticated
-      const url = new URL("/login", request.url);
-      url.searchParams.set("callbackUrl", pathname);
+      const url = new URL('/login', request.url);
+      url.searchParams.set('callbackUrl', pathname);
       return NextResponse.redirect(url);
     }
   }
@@ -28,14 +49,5 @@ export async function middleware(request: NextRequest) {
 }
 
 export const config = {
-  matcher: [
-    /*
-     * Match all request paths except for the ones starting with:
-     * - api (API routes)
-     * - _next/static (static files)
-     * - _next/image (image optimization files)
-     * - favicon.ico (favicon file)
-     */
-    "/((?!api|_next/static|_next/image|favicon.ico).*)",
-  ],
+  matcher: ['/((?!_next/static|_next/image|favicon.ico).*)'],
 };
